Separate search bar scaffold
